Revert fixes for #922707 and #933597 for now.
With the currently implemented fixes, the arch-only build fails.

We cannot move qdoc-qt5 from Build-Depends-Indep to Build-Depends as
that would introduce a dependency loop, so all files generated by qdoc
must be in an Architecture: all package.

A working fix for #922707 would be adding a new qtbase5-doc-dev package,
shipping the .tags file. We will do it after the 5.12 transition, in
order to not delay it further.